fix gradle. reworl geiger view

+++ engine/src/main/kotlin/net/afterday/compas/engine/engine/system/influence/anomaly/AnomalyEvent.kt
@@ -5,7 +5,7 @@ import java.util.*
 class AnomalyEvent (
     val id: String,
     val type: Int,
-    val value: Long, //TODO: need use byte
+    val value: Int, //TODO: need use byte
     val setting: Dictionary<String, String>?,
     val at: Long
 )
